+++ app/src/main/res/layout/item_apps_one.xml
@@ -32,7 +32,7 @@
         tools:text="Title" />
 
     <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/itemOneAppDescriptionTextView"
+        android:id="@+id/itemOneAppRatingTextView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
@@ -45,6 +45,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/itemOneAppTitleTextView"
         app:layout_constraintVertical_bias="0"
-        tools:text="Description" />
+        tools:text="Rating" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
